+++ src/Identity/TokenDto.cs
@@ -1,3 +1,6 @@
 ﻿namespace Light.Identity;
 
-public record TokenDto(string AccessToken, double ExpiresIn, string? RefreshToken);
+public record TokenDto(
+    string AccessToken,
+    long ExpiresIn,
+    string? RefreshToken);
